% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{bayesianSynth}
\alias{bayesianSynth}
\title{Create a bsynth Object Using Panel Data}
\description{
Create a bsynth Object Using Panel Data

Create a bsynth Object Using Panel Data
}
\details{
A Bayesian Synthetic Control has raw data and draws from the posterior
distribution. This is represented by an R6 Class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{draws}}{from the posterior distribution of the lift.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{timeTiles}}{ggplot2 object that shows when
the intervention happened.}

\item{\code{plotData}}{tibble with the observed outcome and the
conterfactual data.}

\item{\code{synthetic}}{ggplot2 object that shows the
observed and conterfactual outcomes over time.}

\item{\code{checks}}{MCMC checks.}

\item{\code{draws}}{from the posterior distribution of the lift.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{bayesianSynth$new()}}
\item \href{#method-fit}{\code{bayesianSynth$fit()}}
\item \href{#method-updateWidth}{\code{bayesianSynth$updateWidth()}}
\item \href{#method-summarizeLift}{\code{bayesianSynth$summarizeLift()}}
\item \href{#method-effect}{\code{bayesianSynth$effect()}}
\item \href{#method-liftDraws}{\code{bayesianSynth$liftDraws()}}
\item \href{#method-clone}{\code{bayesianSynth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new bayesianSynth object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$new(
  data,
  time,
  id,
  treated,
  outcome,
  ci_width = 0.75,
  GP = FALSE,
  covariates = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Long Data frame with outcome, time, id, and treatment
indicator.}

\item{\code{time}}{Name of the variable in the data frame that}

\item{\code{id}}{Name of the variable in the data frame that
contains the standard errors of the point estimates.}

\item{\code{treated}}{Name of the variable in the data frame that contains the
id of the studies.}

\item{\code{outcome}}{Name of the outcome variable.}

\item{\code{ci_width}}{Credible interval's width.  This number in the
(0,1) interval.}

\item{\code{GP}}{Logical that indicates whether or not to include a
Gaussian Process as part of the model.}

\item{\code{covariates}}{Matrix of covariates.
Defaults to NULL if no covariates should be included in the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{bayesianSynth} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit Stan model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$fit(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{other arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-updateWidth"></a>}}
\if{latex}{\out{\hypertarget{method-updateWidth}{}}}
\subsection{Method \code{updateWidth()}}{
Update the width of the credible interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$updateWidth(ci_width = 0.75)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ci_width}}{New width for the credible interval. This number in the
(0,1) interval.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summarizeLift"></a>}}
\if{latex}{\out{\hypertarget{method-summarizeLift}{}}}
\subsection{Method \code{summarizeLift()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$summarizeLift()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-effect"></a>}}
\if{latex}{\out{\hypertarget{method-effect}{}}}
\subsection{Method \code{effect()}}{
effect ggplot2 object that shows the
effect of the intervention over time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$effect(facet = TRUE, subset = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-liftDraws"></a>}}
\if{latex}{\out{\hypertarget{method-liftDraws}{}}}
\subsection{Method \code{liftDraws()}}{
Plots lift.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$liftDraws(from, to, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{First period to consider when calculating lift.}

\item{\code{to}}{Last period to consider when calculating lift.}

\item{\code{...}}{other arguments passed to \code{\link[vizdraws:vizdraws]{vizdraws::vizdraws()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vizdraws object with the posterior distribution of the lift.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bayesianSynth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
