% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsynth-package.R
\docType{package}
\name{bsynth-package}
\alias{bsynth-package}
\alias{bsynth}
\title{The 'bsynth' package.}
\description{
Provides causal inference with a Bayesian synthetic
control method.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
