% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par_long.R
\name{.get_par_long}
\alias{.get_par_long}
\title{Get Parameter Estimates in Long Format}
\usage{
.get_par_long(fit, par)
}
\arguments{
\item{fit}{Stan object with the fitted model.}

\item{par}{Variable to do the long table for.expand_more}
}
\value{
A tibble containing the parameter estimates in long format.
}
\description{
Helper function to get the long dataset of draws given a stan fit object.
}
