% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_tiles.R
\name{time_tiles}
\alias{time_tiles}
\title{Time Tiles Plot of Intervention Impact}
\usage{
time_tiles(data, time, id, status)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{time}{The name of the time period variable (as a string).}

\item{id}{The name of the unit identifier variable (as a string).}

\item{status}{The name of the variable that identifies the treatment status (as a string).}
}
\value{
A ggplot object displaying the time tiles plot.
}
\description{
This function creates a time tiles plot visualizing when and which units are
affected by an intervention. Each tile represents a unit at a specific time point,
with the color indicating the treatment status.
}
